backend/scrapper/requirements.txt
selenium
webdriver_manager
lxml
beautifulsoup4
psycopg2-binary
pandas
